requirements.txt
asyncio
coolname
cryptography
docker
email-validator
flask
flask-babel
flask-login
flask-migrate
flask-sqlalchemy
flask-wtf
gunicorn
humanize
pyjwt
python-dotenv
redis
requests
resend
rq